+++ services/media_engine/plugins/source/http_source/hls/playlist_downloader.cpp
@@ -126,7 +126,7 @@ bool PlayListDownloader::ParseUriInfo(const std::string& uri)
             offset_ = static_cast<int64_t>(fileSize_);
         }
         size_ = static_cast<uint64_t>(std::stoll(fdUriMatch[3].str())); // 3: sub match size subscript
-        uint64_t remainingSize = fileSize_ - offset_;
+        uint64_t remainingSize = fileSize_ - static_cast<uint64_t>(offset_);
         if (size_ > remainingSize) {
             size_ = remainingSize;
         }
